[Gtk] Check for missing python modules and command line programs
Add stuff to check that the appropriate python modules and commands are installed, checks for Xlib, dbus, pynotify and gi, along with other python modules, that in my understanding should be distributed with most versions of python, but better safe than sorry. If any packages are missing the program will quit with `sys.exit`.

Checks if programs are installed as well using `which` from `shutil`. Only checks for currently required programs to be installed as dictated for autokey-gtk in debian/control, (zenity, wmctrl, ps).